Fix some types at the IPC layer. Sizes should be unsigned.
BUG=164958
Review URL: https://codereview.chromium.org/11474040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@172474 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b31199ef64e72567c9f66e49d7734e361357e375

+++ webkit/plugins/ppapi/ppb_video_decoder_impl.cc
@@ -142,7 +142,7 @@ int32_t PPB_VideoDecoder_Impl::Decode(
   media::BitstreamBuffer decode_buffer(
       bitstream_buffer->id,
       buffer->shared_memory()->handle(),
-      static_cast<size_t>(bitstream_buffer->size));
+      bitstream_buffer->size);
   if (!SetBitstreamBufferCallback(bitstream_buffer->id, callback))
     return PP_ERROR_BADARGUMENT;
